Further networking refactor - made NetworkDataModel base data class
And changed the old NetworkDataModel to now be MobileSignalNetworkDataModel which just holds
Mobile Signal info; this lets us reuse the existing methods for sending various implementations
(subclasses) of NetworkDataModel without much hassle; and only one instantiation of old NetworkDataModel
had to be changed to now instantiate the new MobileSignalNetworkDataModel

So Iperf3 can start sending a new subclass of NetworkDataModel e.g. Iperf3NetworkDataModel to send it's stuff
